requirements.txt
Pillow==9.5.0
rich==13.4.1
webcolors==1.13